added vdp pointer to tile-based resources
converted Map to be loaded through the resource manager
removed the vdp location to level002 spawn data

+++ project/source/engine/scroll_manager.c
@@ -17,7 +17,7 @@ u16 ScrollManager_mapHeight;
 const u16* ScrollManager_metatileMap;
 const u16* ScrollManager_metatileLut;
 
-u8 ScrollManager_tilesetVdpLocation;
+u16 ScrollManager_tilesetVdpLocation;
 
 
 u8 ScrollManager_speedX = 0;
