+++ project/source/engine/scroll_manager.h
@@ -32,6 +32,6 @@ extern u16 ScrollManager_mapHeight;
 extern const u16* ScrollManager_metatileMap;
 extern const u16* ScrollManager_metatileLut;
 
-extern u8 ScrollManager_tilesetVdpLocation;
+extern u16 ScrollManager_tilesetVdpLocation;
 
 #endif
